requirements.txt
fastapi
uvicorn
ultralytics
pymongo
numpy
pillow
python-multipart
